# Optimized Enum::byOrdinal() and Enum::getName[s]() 